Add signer in constructor
